Send SMS over 1x when on eHRPD and active voice call
When UE is attached to eHRPD network and IMS is registered:
- If there is an active 1x RTT voice call, send SMS over 1x
- If there no active 1xRTT voice call, send SMS over IMS
This feature is enabled by default. To disable it, set
config_send_sms1x_on_voice_call to false in config.xml

- If MO SMS fails with SMS_FAIL_RETRY and phone's radio tech is CDMA,
 send the subsequent retry sms on 1x if above criteria for active 1x
 voice call and IMS registration still apply

- Revert AOSP commit 4eb8e4c8dc684e2109276ea2ec4ce25d3b188568.
The functionality implemented by this change is already there in our internal
change (e87a8e9e7cd6f645a051b5446f27eb547a4de636)
The problem with this change is that it assumes that for a 3gpp2 operator,
the sms format will always be 3gpp2 even on IMS which may not be so.
In such a case, this change won't work. Also, this change has one additional
conditional check for LTE + no concurrent voice and data allowed which is taken
care of by our modem

Change-Id: Ic953299e5f1db518a3268c5203d51dffbe370408

Change-Id: I797b9218ef013402fae2d01242f1f7b9a02c64f2

CRs-Fixed: 565247, 607158, 750718

Change-Id: I767ff1747c8498a5fd3760e2a8a6cbd764716b37