Fixes for loading SIM contact in CDMA.
Unable to load sim contacts when in cdma network mode.

1)Translate EF_ADN to EF_PBR if the current app is CSIM:

If the current app is CSIM, translate EF_ADN to EF_PBR while
loading the sim contacts.

2)Update the IccRecords in phonebook when icc status changes.

For cdma phone, currently IccRecords in the phone book are not
are being updated when the sim state update is received. Due to this
the records in the phonebook are null and hence the adn cache in the
phonebook is also null.

When the icc state changes, update the icc records in the phone.

Change-Id: I33034a5a106286e3f63df0fb7201b20ccdc2aa40